Debugged the full code.

+++ backend/src/main/java/com/nsbm_projects/hotel_management_system/dto/RoomResponse.java
@@ -3,12 +3,15 @@ package com.nsbm_projects.hotel_management_system.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class RoomResponse {
+
     private Long id;
     private String roomNumber;
     private String type;
-    private double pricePerNight;
+    private BigDecimal pricePerNight;   // ✅ FIXED
     private boolean available;
 }
